requirements.txt
flask>=3.1.2
flask-cors>=6.0.1
flask-sqlalchemy>=3.1.0
gunicorn>=23.0.0
psycopg2-binary>=2.9.11
python-dotenv>=1.1.1
requests>=2.31.0
email_validator>=2.1.0
pyTelegramBotAPI>=4.14.0
cloudinary>=1.36.0
werkzeug>=3.0.0
cryptography>=41.0.0
google-generativeai>=0.3.0
